Add bulk expire for silences
Adds the ability to bulk expire silences from the silences list page.

Adds a checkbox per row to the silences list for selecting silences as
well as a checkbox for selecting / unselecting all of the silences.

If the list is currently filtered, the select all checkbox only selects
the currently displayed silences, but unselecting all also unselects any
silences that are currently hidden.