Disable select all silences checkbox if all silences are expired
If all of the currently visible silences are already expired, disable
the checkbox for selecting all silences.

This is better UX because we disable each silence's individual checkbox
if it is expired and we should be consistent with that. If in the future
we add a bulk action for expired silences, we can change this behavior.