Fix bug where the variable dropdown options were not updated
Changing the selected value for one variable dropdown can cause the
available options for other variable dropdowns to be updated. In this
case, those options would sometimes fail to update.

The bug seems to be caused by the PatternFly <Select /> not rerendering
the component under certain circumstances. Rendering the select options
inline works around the problem.