targets: filter by target
Previously, a user could only filter by the endpoint or namespace.
Endpoints are generally the same between components and not useful for
filtering, as well as being opaque. Namespaces are useful but not
precise. Filtering by the target name (PodMonitor or ServiceMonitor)
will be more useful for narrowing down the list.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>