On the alert details page, don't require all alert labels in URL
Before this change we required all of the alert's labels to be in the
URL query params. This change allows the alert to be shown even if only
some of its labels are in the URL, as long as there is only one such
matching alert.

Fixes a bug where the alerts links in "Notifications" didn't work
because they didn't include the alert's external labels.