fix(cypress): Update Incident Traversal
The UI changes introduced few quirks that break the traversal of
incidents, but also solved some that had existing workarounds in the
testsuite.

This commit removes the retry for the not loading alerts chart, as it is
loading reliably now, but because stale state is loaded first, it
introduces hard coded wait time.

Additionally, it fixes the failing traversal when no incidents exits in
the UI and adds test case that verifies that on mocked data.