OCPBUGS-66064: use max TLS version only when defined
This commit ensures that the TLS max version is set only when explicitly
configured from the command-line (or environment variable). In the
previous version, the binary always defaulted to TLS Version 1.2 and it
created an issue with the "modern" TLS profile which defines 1.3 as the
minimum TLS (e.g. min version > max version).

Signed-off-by: Simon Pasquier <spasquie@redhat.com>